Refactor btree/node/mod.rs: Move tests to separate file
Extract all test code from mod.rs (955 lines) into tests.rs (924 lines).
The main module file is now only 30 lines, improving readability and
maintainability while keeping all tests functional.

- Created tests.rs with all 34 test functions
- Updated mod.rs to reference tests module
- All tests pass (verified with cargo test)
- No functionality changes

Closes #1913
